refactor: put layout components together
