+++ src/Widget/EnglishCard/EnglishCard.tsx
@@ -5,7 +5,7 @@ import 'swiper/css';
 import styled from 'styled-components';
 import { getCard } from '../../utils/firebase';
 import EnglishWord, { EnglishWordData } from './EnglishWord';
-import { SwiperButtonNext, SwiperButtonPrev } from '../../Swiper';
+import { SwiperButtonNext, SwiperButtonPrev } from '../../components/Swiper';
 import { DoubleArrow, Refresh } from '../../components/Icons';
 import useLocalStorage from '../../hooks/useLocalStorage';
 import { handleNewEnglishWords, afterOneDay } from './utils';
