+++ src/components/Swiper/index.ts
