+++ src/layout/NavBar/SelectNewWidget/index.ts
@@ -1,0 +1,3 @@
+import SelectNewWidget from './SelectNewWidget';
+
+export default SelectNewWidget;
