+++ src/layout/NavBar/ShortCuts/AddNewShortCut.tsx
@@ -1,8 +1,8 @@
 import React, { useRef, useState } from 'react';
 import styled, { css } from 'styled-components';
-import FormInput from '../components/FormInput';
-import Button from '../components/Button/Button';
-import { getValidateURL } from '../utils/lib';
+import FormInput from '../../../components/FormInput';
+import Button from '../../../components/Button/Button';
+import { getValidateURL } from '../../../utils/lib';
 
 const Wrapper = styled.div`
   position: absolute;
