+++ src/layout/NavBar/ShortCuts/defaultShortCut.ts
@@ -1,0 +1,19 @@
+const defaultShortCuts = [
+  {
+    id: '1',
+    title: 'Netflix',
+    url: 'https://www.netflix.com/browse',
+  },
+  {
+    id: '2',
+    title: 'Google',
+    url: 'https://www.google.com.tw/',
+  },
+  {
+    id: '3',
+    title: 'Youtube',
+    url: 'http://www.youtube.com/?hl=zh-TW&gl=TW',
+  },
+] as const;
+
+export default defaultShortCuts;
