+++ src/layout/NavBar/index.ts
@@ -1,0 +1,3 @@
+import NavBar from './NavBar';
+
+export default NavBar;
